requirements.txt
django>=4.2
psycopg2-binary>=2.9
python-dotenv>=1.0.0
openpyxl>=3.1.2
Pillow>=10.0.0
gunicorn>=21.0

